requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --allow-unsafe
#
arrow==0.13.2             # via cookiecutter-python-project (setup.py), jinja2-time
binaryornot==0.4.4        # via cookiecutter
certifi==2020.6.20        # via requests
cffi==1.14.3              # via pynacl
chardet==3.0.4            # via binaryornot, requests
click==7.1.2              # via cookiecutter, passpy
cookiecutter==1.7.2       # via cookiecutter-python-project (setup.py)
gitdb==4.0.5              # via gitpython
gitpython==3.1.9          # via passpy
idna==2.10                # via requests
jinja2-time==0.2.0        # via cookiecutter
jinja2==2.11.2            # via cookiecutter, jinja2-time
markupsafe==1.1.1         # via cookiecutter, jinja2
passpy==1.0.2             # via cookiecutter-python-project (setup.py)
poyo==0.5.0               # via cookiecutter
pycparser==2.20           # via cffi
pynacl==1.4.0             # via cookiecutter-python-project (setup.py)
pyperclip==1.8.1          # via passpy
python-dateutil==2.8.1    # via arrow
python-gnupg==0.4.6       # via passpy
python-slugify==4.0.1     # via cookiecutter
requests==2.24.0          # via cookiecutter
sh==1.14.0                # via cookiecutter-python-project (setup.py)
six==1.15.0               # via cookiecutter, pynacl, python-dateutil
smmap==3.0.4              # via gitdb
text-unidecode==1.3       # via python-slugify
urllib3==1.25.10          # via requests
